+++ src/utils/getModifierClass.ts
@@ -1,9 +1,9 @@
-import { BemProps, Modifiers } from '../types';
+import { Modifiers } from '../types';
 
 const getModifierClass = (
 	/** BEM prefix */
 	bem: string,
-	modifier: BemProps['modifier']
+	modifier: Modifiers
 ): string => {
 	const flattenModifiers = (modifier: Modifiers): string =>
 		!modifier || !modifier.length
@@ -15,4 +15,5 @@ const getModifierClass = (
 	return flattenModifiers(modifier);
 };
 
+/** @deprecated Instead `import { modifiedClass } from '@hugsmidjan/qj/classUtils';`  (Will be removed in v0.5) */
 export default getModifierClass;